requirements.txt
pylint
twitter
discord.py
discord.py-stubs
